OSIS-1330 Rename group.py to create.py
	renommé :         base/forms/education_group/group.py -> base/forms/education_group/create.py
	modifié :         base/tests/forms/education_group/test_create.py
	modifié :         base/tests/views/education_groups/test_create.py
	modifié :         base/tests/views/education_groups/test_update.py
	modifié :         base/views/education_groups/create.py
	modifié :         base/views/education_groups/update.py
